requirements.txt
Django
South
django-south-admin
django-annoying
gunicorn
wsgiref
tornado
usethis-django-bootstrap
psycopg2
autopep8
django-admin-bootstrapped
dj-database-url
